Adicionado proteções de permissões

+++ playxchangewebapp/backend/controllers/FranquiaController.php
@@ -36,6 +36,10 @@ class FranquiaController extends Controller
                         'roles' => ['admin','funcionario','moderador'],
                     ],
                 ],
+                'denyCallback' => function () {
+                    \Yii::$app->session->setFlash('error', 'Não possui permissões suficientes para executar esta ação!');
+                    $this->goHome();
+                }
             ],
             'verbs' => [
                 'class' => VerbFilter::class,
@@ -52,14 +56,17 @@ class FranquiaController extends Controller
      */
     public function actionIndex()
     {
+        if(Yii::$app->user->can('verTudo')) {
+            $searchModel = new FranquiaSearch();
+            $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
-        $searchModel = new FranquiaSearch();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
-
-        return $this->render('index', [
-            'dataProvider' => $dataProvider,
-            'searchModel' => $searchModel,
-        ]);
+            return $this->render('index', [
+                'dataProvider' => $dataProvider,
+                'searchModel' => $searchModel,
+            ]);
+        }else{
+            return $this->goHome();
+        }
     }
 
     /**
@@ -70,9 +77,14 @@ class FranquiaController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if(Yii::$app->user->can('verDetalhesFranquias')){
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        }else{
+            return $this->goHome();
+        }
+
     }
 
     /**
@@ -82,15 +94,20 @@ class FranquiaController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Franquia();
+        if(Yii::$app->user->can('adicionarFranquias')) {
+            $model = new Franquia();
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+
+            return $this->render('create', [
+                'model' => $model,
+            ]);
+        }else{
+            return $this->goHome();
         }
 
-        return $this->render('create', [
-            'model' => $model,
-        ]);
     }
 
     /**
@@ -102,15 +119,20 @@ class FranquiaController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if(Yii::$app->user->can('editarFranquias')){
+            $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+
+            return $this->render('update', [
+                'model' => $model,
+            ]);
+        }else{
+            return $this->goHome();
         }
 
-        return $this->render('update', [
-            'model' => $model,
-        ]);
     }
 
     /**
@@ -122,9 +144,13 @@ class FranquiaController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if(Yii::$app->user->can('removerFranquias')){
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        }else{
+            return $this->goHome();
+        }
     }
 
     /**
